move stuff to reqs folder
