+++ src/Domain.LinnApps/Requisitions/StoresTransactionDefinition.cs
@@ -1,4 +1,4 @@
-﻿namespace Linn.Stores2.Domain.LinnApps
+﻿namespace Linn.Stores2.Domain.LinnApps.Requisitions
 {
     public class StoresTransactionDefinition
     {
