+++ src/Facade/ResourceBuilders/StoresFunctionCodeResourceBuilder.cs
@@ -3,6 +3,7 @@
     using System.Collections.Generic;
     using Linn.Common.Facade;
     using Linn.Stores2.Domain.LinnApps;
+    using Linn.Stores2.Domain.LinnApps.Requisitions;
     using Linn.Stores2.Resources.Requisitions;
 
     public class StoresFunctionCodeResourceBuilder : IBuilder<StoresFunctionCode>
