+++ tests/Integration/Integration.Tests/RequisitionModuleTests/WhenGettingFunctionCodes.cs
@@ -6,6 +6,7 @@
 
     using FluentAssertions;
     using Linn.Stores2.Domain.LinnApps;
+    using Linn.Stores2.Domain.LinnApps.Requisitions;
     using Linn.Stores2.Integration.Tests.Extensions;
     using Linn.Stores2.Resources.Requisitions;
