+++ tests/Unit/Domain.LinnApps.Tests/RequisitionHeaderTests/WhenCancellingLineAndBooked.cs
@@ -5,6 +5,7 @@
     using FluentAssertions;
 
     using Linn.Stores2.Domain.LinnApps.Exceptions;
+    using Linn.Stores2.Domain.LinnApps.Requisitions;
     using Linn.Stores2.TestData.Requisitions;
 
     using NUnit.Framework;
